requirements.txt
absl-py==0.13.0
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.3.1
astunparse==1.6.3
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.2
celery==5.0.5
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-cors-headers==3.8.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
fcm-django==1.0.5
firebase-admin==5.0.2
flatbuffers==1.12
gast==0.4.0
google-api-core==1.31.2
google-api-python-client==2.19.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kombu==5.1.0
login==0.0.6
Markdown==3.3.4
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.1.2
promise==2.3
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
